% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dados_sorteios}
\alias{dados_sorteios}
\title{Le o conjunto de dados com todos os sorteios até a data atual
de uma modalidade de loteria}
\usage{
dados_sorteios(modalidade)
}
\arguments{
\item{modalidade}{modalidade do sorteio, disponiveis:
\itemize{
\item megasena
\item lotofacil
\item quina
\item lotomania
\item duplasena
\item diadesorte
\item supersete
}}
}
\value{
uma \code{tibble}
}
\description{
Le o conjunto de dados com todos os sorteios até a data atual
de uma modalidade de loteria
}
\examples{
dados_sorteios(modalidade = 'megasena')
}
