% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diadesorte}
\alias{diadesorte}
\title{Dia de Sorte.}
\format{
Um conjunto de dados com 9 variaveis: \code{data}, \code{concurso},
\code{mes}, \code{dezena_1}, \code{dezena_2}, \code{dezena_3}, \code{dezena_4},
\code{dezena_5}, \code{dezena_6} e \code{dezena_7}.
}
\usage{
diadesorte
}
\description{
Todos os sorteios da modalidade "Dia de Sorte" das Loterias Caixa
}
\keyword{datasets}
