% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{duplasena}
\alias{duplasena}
\title{Dupla Sena.}
\format{
Um conjunto de dados com 14 variaveis: \code{date}, \code{id_concurso},
\code{sort_1}, \code{sort_2}, \code{sort_3}, \code{sort_4}, \code{sort_5},
\code{sort_6}, \code{sort_7}, \code{sort_8}, \code{sort_9}, \code{sort_10},
\code{sort_11} e \code{sort_12}.
}
\usage{
duplasena
}
\description{
Todos os sorteios da modalidade "Dupla Sena" das Loterias Caixa
}
\keyword{datasets}
