% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotomania}
\alias{lotomania}
\title{Lotomania.}
\format{
Um conjunto de dados com 22 variaveis: \code{data}, \code{concurso},
\code{dezena_1}, \code{dezena_2}, \code{dezena_3}, \code{dezena_4}, \code{dezena_5},
\code{dezena_6}, \code{dezena_7}, \code{dezena_8}, \code{dezena_9}, \code{dezena_10},
\code{dezena_11}, \code{dezena_12}, \code{dezena_13}, \code{dezena_14}, \code{dezena_15},
\code{dezena_16}, \code{dezena_17}, \code{dezena_18}, \code{dezena_19} e \code{dezena_20}.
}
\usage{
lotomania
}
\description{
Todos os sorteios da modalidade "Lotomania" das Loterias Caixa
}
\keyword{datasets}
