% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{megasena}
\alias{megasena}
\title{Mega-Sena.}
\format{
Um conjunto de dados com 8 variaveis: \code{data}, \code{concurso},
\code{dezena_1}, \code{dezena_2}, \code{dezena_3}, \code{dezena_4}, \code{dezena_5} e
\code{dezena_6}.
}
\usage{
megasena
}
\description{
Todos os sorteios da modalidade "Mega-Sena" das Loterias Caixa
}
\keyword{datasets}
