% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{necessario_atualizar}
\alias{necessario_atualizar}
\title{Verifica se ha novos sorteios a serem coletados no dataset
historico da modalidade de loteria}
\usage{
necessario_atualizar(modalidade)
}
\arguments{
\item{modalidade}{modalidade do sorteio, disponiveis:
\itemize{
\item megasena
\item lotofacil
\item quina
\item lotomania
\item duplasena
\item diadesorte
\item supersete
}}
}
\value{
um \code{logico}
}
\description{
Verifica se ha novos sorteios a serem coletados no dataset
historico da modalidade de loteria
}
\examples{
necessario_atualizar(modalidade = 'megasena')
}
