% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obter_json_sorteio}
\alias{obter_json_sorteio}
\title{Obtem o JSON de um sorteio especifico de uma
modalidade de loteria}
\usage{
obter_json_sorteio(id_concurso, modalidade)
}
\arguments{
\item{id_concurso}{numero do concurso do sorteio}

\item{modalidade}{modalidade do sorteio, disponiveis:
\itemize{
\item megasena
\item lotofacil
\item quina
\item lotomania
\item duplasena
\item diadesorte
\item supersete
}}
}
\value{
uma \code{string}
}
\description{
Obtem o JSON de um sorteio especifico de uma
modalidade de loteria
}
\examples{
obter_json_sorteio(id_concurso = 1, modalidade = 'megasena')
}
