% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quina}
\alias{quina}
\title{Quina.}
\format{
Um conjunto de dados com 7 variaveis: \code{data}, \code{concurso},
\code{dezena_1}, \code{dezena_2}, \code{dezena_3}, \code{dezena_4} e \code{dezena_5}.
}
\usage{
quina
}
\description{
Todos os sorteios da modalidade "Quina" das Loterias Caixa
}
\keyword{datasets}
