% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resultado_loteria}
\alias{resultado_loteria}
\title{Obtem o resultado de um sorteio especifico de uma
modalidade de loteria}
\usage{
resultado_loteria(concurso = NULL, modalidade)
}
\arguments{
\item{concurso}{numero do concurso do sorteio}

\item{modalidade}{modalidade do sorteio, disponiveis:
\itemize{
\item megasena
\item lotofacil
\item quina
\item lotomania
\item duplasena
\item diadesorte
\item supersete
}}
}
\value{
uma \code{tibble}
}
\description{
Obtem o resultado de um sorteio especifico de uma
modalidade de loteria
}
\examples{
resultado_loteria(concurso = 1, modalidade = 'megasena')
}
