% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resultado_loteria_todos}
\alias{resultado_loteria_todos}
\title{Obtem o resultado de todos os sorteios até a data atual
de uma modalidade de loteria, ou partindo de um concurso}
\usage{
resultado_loteria_todos(modalidade, min_concurso = 1, max_concurso = NULL)
}
\arguments{
\item{modalidade}{modalidade do sorteio, disponiveis:
\itemize{
\item megasena
\item lotofacil
\item quina
\item lotomania
\item duplasena
\item diadesorte
\item supersete
}}

\item{min_concurso}{numero do concurso de partida, padrao: 1}

\item{max_concurso}{numero maximo do concurso, padrao: ultimo sorteio}
}
\value{
uma \code{tibble}
}
\description{
Obtem o resultado de todos os sorteios até a data atual
de uma modalidade de loteria, ou partindo de um concurso
}
\examples{
resultado_loteria_todos(modalidade = 'megasena', min_concurso = 2452)
}
