% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supersete}
\alias{supersete}
\title{Super Sete.}
\format{
Um conjunto de dados com 9 variaveis: \code{date}, \code{id_concurso},
\code{sort_1}, \code{sort_2}, \code{sort_3}, \code{sort_4}, \code{sort_5},
\code{sort_6} e\code{sort_7}.
}
\usage{
supersete
}
\description{
Todos os sorteios da modalidade "Super Sete" das Loterias Caixa
}
\keyword{datasets}
