% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timemania}
\alias{timemania}
\title{Timemania.}
\format{
Um conjunto de dados com 22 variaveis: \code{data}, \code{concurso},
\code{time}, \code{dezena_1}, \code{dezena_2}, \code{dezena_3}, \code{dezena_4},
\code{dezena_5}, \code{dezena_6} e \code{dezena_7}.
}
\usage{
timemania
}
\description{
Todos os sorteios da modalidade "Timemania" das Loterias Caixa
}
\keyword{datasets}
