% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diadesorte}
\alias{diadesorte}
\title{Dia de Sorte.}
\format{
Um conjunto de dados com 9 variaveis: \code{date}, \code{id_concurso},
\code{sort_1}, \code{sort_2}, \code{sort_3}, \code{sort_4}, \code{sort_5},
\code{sort_6} e\code{sort_7}.
}
\usage{
diadesorte
}
\description{
Todos os sorteios da modalidade "Dia de Sorte" das Loterias Caixa
}
\keyword{datasets}
