% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotomania}
\alias{lotomania}
\title{Lotomania.}
\format{
Um conjunto de dados com 22 variaveis: \code{date}, \code{id_concurso},
\code{sort_1}, \code{sort_2}, \code{sort_3}, \code{sort_4}, \code{sort_5},
\code{sort_6}, \code{sort_7}, \code{sort_8}, \code{sort_9}, \code{sort_10},
\code{sort_11}, \code{sort_12}, \code{sort_13}, \code{sort_14}, \code{sort_15},
\code{sort_16}, \code{sort_17}, \code{sort_18}, \code{sort_19} e \code{sort_20}.
}
\usage{
lotomania
}
\description{
Todos os sorteios da modalidade "Lotomania" das Loterias Caixa
}
\keyword{datasets}
