% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{megasena}
\alias{megasena}
\title{Mega-Sena.}
\format{
Um conjunto de dados com 8 variaveis: \code{date}, \code{id_concurso},
\code{sort_1}, \code{sort_2}, \code{sort_3}, \code{sort_4}, \code{sort_5} e
\code{sort_6}.
}
\usage{
megasena
}
\description{
Todos os sorteios da modalidade "Mega-Sena" das Loterias Caixa
}
\keyword{datasets}
