% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quina}
\alias{quina}
\title{Quina.}
\format{
Um conjunto de dados com 7 variaveis: \code{date}, \code{id_concurso},
\code{sort_1}, \code{sort_2}, \code{sort_3}, \code{sort_4} e \code{sort_5}.
}
\usage{
quina
}
\description{
Todos os sorteios da modalidade "Quina" das Loterias Caixa
}
\keyword{datasets}
