% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supersete}
\alias{supersete}
\title{Super Sete.}
\format{
Um conjunto de dados com 9 variaveis: \code{data}, \code{concurso},
\code{dezena_1}, \code{dezena_2}, \code{dezena_3}, \code{dezena_4}, \code{dezena_5},
\code{dezena_6} e\code{dezena_7}.
}
\usage{
supersete
}
\description{
Todos os sorteios da modalidade "Super Sete" das Loterias Caixa
}
\keyword{datasets}
